Tags, Game Over function, Movement fixes
Added "Ball" tag to the balls, "Player" tag to the player, and "Top" to TopBorder. Added Game Over text and functionality.
Lifted the player and fixed a bug where the ball that was being carried by the player would not follow the player's movement when the player was moving left.